+++ RoguelikeFramework/Assets/Scripts/Effects/Core/Stun.cs
@@ -81,7 +81,7 @@ public class Stun : PipEffect
     //public override void RegenerateStats(ref Stats stats) {}
 
     //Called wenever a monster gains energy
-    public override void OnEnergyGained(ref int energy)
+    public override void OnEnergyGained(ref float energy)
     {
         energy = 0;
     }
